refactor cloudflare worker with client stream
